Almost done???

+++ Campaign.cs
@@ -11,10 +11,10 @@ namespace Kassasystem
         public string CampaignName { get; set; }
         public DateTime CampaignStartDate { get; set; }
         public DateTime CampaignEndDate { get; set; }
-        public float CampaignDiscountPercent { get; set; }
+        public short CampaignDiscountPercent { get; set; }
         public List<Product> ProductsInCampaign { get; set; }
 
-        public Campaign(string name, DateTime start, DateTime end, float discount, List<Product> productsInCampaign)
+        public Campaign(string name, DateTime start, DateTime end, short discount, List<Product> productsInCampaign)
         {
             CampaignName = name;
             CampaignStartDate = start;
